NewSeg() is starting to be replaced with addSeg(Segment seg) which will help with when we make Segment immutable.
BasicFile now has an addSeg(Segment seg) method that just adds the segment to the end of the active segs list. This currently throws an exception if the fileid/filename/removed fields are not correct, although this is bad programming. Later will change the design so Segment never even has these fields. 

Added the BasicFileException test to check that the IllegalArgumentException is thrown properly (not a real JUnit test, cuz I was too lazy to figure out JUnit exception testing).